# contracts/http/api-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: CreatorFlow API Gateway
  version: 0.1.0
  description: Public API gateway untuk MVP CreatorFlow.
servers:
  - url: https://api.creatorflow.local
...
paths:
...
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user and workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /v1/me:
    get:
      tags: [Users]
      summary: Get current user and workspace context
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current session context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, workspaceName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
        workspaceName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [userId, workspaceId, accessToken, expiresIn]
      properties:
        userId:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        accessToken:
          type: string
        expiresIn:
          type: string
    MeResponse:
      type: object
      required: [user, workspace]
      properties:
        user:
          type: object
          required: [id, email, status]
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            fullName:
              type: string
              nullable: true
            status:
              type: string
        workspace:
          type: object
          required: [id, name, roleCode, slug]
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            roleCode:
              type: string
            slug:
              type: string
...